+++ gsa/src/web/graphql/tasks.js
@@ -501,6 +501,7 @@ export const useExportTasksByFilter = options => {
     options,
   );
   const exportTasksByFilter = useCallback(
+    // eslint-disable-next-line no-shadow
     filterString =>
       queryExportTasksByFilter({
         ...options,
@@ -508,7 +509,7 @@ export const useExportTasksByFilter = options => {
           filterString,
         },
       }),
-    [queryExportTasksByFilter],
+    [queryExportTasksByFilter, options],
   );
 
   const exportedTasks = data?.exportFilteredTasks?.exportedEntities.substring(
